Refactor UserFactory
