+++ .gitignore
@@ -18,3 +18,4 @@ yarn-error.log
 /_ide_helper.php
 .DS_Store
 .php-cs-fixer.cache
+.phpstorm.meta.php
