+++ database/factories/UserFactory.php
@@ -3,33 +3,27 @@
 namespace Database\Factories;
 
 use Domain\User\Models\User;
-use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
 
-/**
- * @extends Factory<User>
- */
-class UserFactory extends Factory {
-  /**
-   * The name of the factory's corresponding model.
-   *
-   * @unreleased
-   */
-  protected $model = User::class;
+class UserFactory {
+  public Carbon|int|null $email_verified_at = 0;
 
-  /**
-   * Define the model's default state.
-   *
-   * @unreleased
-   */
-  public function definition(): array {
-    return [
+  public static function new(): self {
+    return new self();
+  }
+
+  public function create(array $overrides = []): User {
+    $user = User::create(array_merge([
       'name' => fake()->name(),
       'email' => fake()->safeEmail(),
-      'email_verified_at' => now(),
       'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
-      'remember_token' => Str::random(10),
-    ];
+    ], $overrides));
+
+    $user->email_verified_at = $this->email_verified_at === 0 ? now() : $this->email_verified_at;
+    $user->remember_token = Str::random(10);
+
+    return $user;
   }
 
   /**
@@ -38,10 +32,8 @@ class UserFactory extends Factory {
    * @unreleased
    */
   public function unverified(): static {
-    return $this->state(function (array $attributes) {
-      return [
-        'email_verified_at' => null,
-      ];
-    });
+    $clone = clone $this;
+    $clone->email_verified_at = null;
+    return $clone;
   }
 }
